move test to test_coco_evaluation.py
